fea: add binaries for electron_v31

+++ src/odbc.cpp
@@ -1599,7 +1599,7 @@ Local<Value> ODBC::GetSQLError (SQLSMALLINT handleType, SQLHANDLE handle, char* 
       DEBUG_TPRINTF(SQL_T("ODBC::GetSQLError : errorMessage=%s, errorSQLState=%s\n"), errorMessage, errorSQLState);
 
       Nan::Set(objError, Nan::New("error").ToLocalChecked(), Nan::New(message).ToLocalChecked());
-      Nan::Set(objError, Nan::New("sqlcode").ToLocalChecked(), Nan::New(native));
+      Nan::Set(objError, Nan::New("sqlcode").ToLocalChecked(), Nan::New((double)native));
 #ifdef UNICODE
       Nan::SetPrototype(objError, Exception::Error(Nan::New((uint16_t *) errorMessage).ToLocalChecked()));
       Nan::Set(objError, Nan::New("message").ToLocalChecked(), Nan::New((uint16_t *) errorMessage).ToLocalChecked());
